+++ src/App.js
@@ -26,8 +26,6 @@ import ProtectedRoute from "./components/ProtectedRoute";
 import Profile from "./pages/Profile";
 import NewTinerary from "./pages/NewTinerary";
 import NewShow from "./pages/NewShow.jsx";
-import Comments from "./components/Comments";
-
 import NewReaction from "./pages/NewReaction.jsx";
 import MyReactions from "./pages/MyReactions";
 
@@ -60,7 +58,6 @@ function App() {
         <Route path="/signin" element={<SignIn />} />
 
 
-        <Route path="/comment" element={<Comments />} />
 
         {/* <Route path="/myreactions" element={<MyReactions />} /> */}
